BUG: Adjust to new style.css loading pattern for Hemingway.
In update 1.57 of Hemingway the author decided to change the way the style.css
was loaded and that lead to our own srtyle.css was loaded before the parent and
no modifications would stick.

Defining the hemingway_load_style() function and extending it with a correct
load order where our style.css is loaded last.